add statistics & advantages sections
