--- src/components/Advantages/index.ts
@@ -1,1 +1,0 @@
-export * from './Advantages';
