--- src/components/Advantages/types.ts
@@ -1,3 +1,0 @@
-export interface AdvantagesProps {
-  className?: string;
-}
