--- src/components/Statistics/index.ts
@@ -1,1 +1,0 @@
-export * from './Statistics';
