+++ src/components/ui/Parallax/Parallax.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import {
   motion,
   useScroll,
