+++ src/views/StatisticsSection/types.ts
